Add warning dialog to deletions

+++ app/containers/VisibleDvds.js
@@ -3,6 +3,8 @@ import { deleteDvd, displayDvd, editDvd } from '../actions'
 import { openSidebar } from '../mixins/ToggleSidebar'
 import DvdList from '../components/DvdList'
 
+const { dialog } = window.require('electron').remote
+
 const mapStateToProps = (state) => {
   return {
     dvds: state.dvds,
@@ -13,7 +15,10 @@ const mapStateToProps = (state) => {
 const mapDispatchToProps = (dispatch) => {
   return {
     onDestroy: (id) => {
-      dispatch(deleteDvd(id))
+      var answer = dialog.showMessageBox({type: 'warning', buttons: ["OK", "Cancel"], title: "Confirm Delete", message: "Are you sure you want to delete this DVD?"})
+      if(answer == 0) {
+        dispatch(deleteDvd(id))
+      }
     },
     displayDvd: (id) => {
       dispatch(displayDvd(id))
